Fix CI clippy and doc failures
- Remove unnecessary `as u64` cast (meta.size is already u64 in 0.13.2)
- Allow clippy::result_large_err on execute_read_ranges (object_store::Error
  is large by design)
- Fix broken rustdoc links to LocalFileSystem (cfg-gated away when
  io-uring feature is enabled)

+++ datafusion/object-store-iouring/src/lib.rs
@@ -169,7 +169,7 @@ impl IoUringObjectStore {
     ) -> Result<GetResult> {
         // Get file metadata via the inner store
         let meta = self.inner.head(location).await?;
-        let file_size = meta.size as u64;
+        let file_size = meta.size;
 
         // Resolve the requested byte range
         let range = match &options.range {
